Move ITs back into a single class
